refactored home page into new file and moved avenger_detail into new folder for best coding practices
