+++ .idea/hacktoberfest21-flutter-gdsc-iiitv.iml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module type="JAVA_MODULE" version="4">
+  <component name="NewModuleRootManager" inherit-compiler-output="true">
+    <exclude-output />
+    <content url="file://$MODULE_DIR$">
+      <excludeFolder url="file://$MODULE_DIR$/hacktoberfest21_gdsc/.dart_tool" />
+      <excludeFolder url="file://$MODULE_DIR$/hacktoberfest21_gdsc/build" />
+      <excludeFolder url="file://$MODULE_DIR$/hacktoberfest21_gdsc/.pub" />
+    </content>
+    <orderEntry type="inheritedJdk" />
+    <orderEntry type="sourceFolder" forTests="false" />
+    <orderEntry type="library" name="Dart SDK" level="project" />
+  </component>
+</module>
